Bugfix: handle tasks without budget in update_enforcement_timer()
This patch addresses a bug reported by Jeremy Erickson. The BUG_ON()
statement in arm_enforcement_timer() could be triggered under GSN-EDF
if tasks exhaust their budget while in the ready queue. The correct
action to take if presented with a budget-less task is to reschedule
immediately (without setting a timer first).